+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @winrlabs/api
 
+## 1.1.0
+
+### Minor Changes
+
+- af62052: Constant url
+
 ## 1.0.1
 
 ### Patch Changes