+++ packages/web3-games/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Minor Changes
 
+- af62052: Constant url
+
+### Patch Changes
+
+- Updated dependencies [af62052]
+  - @winrlabs/api@1.1.0
+  - @winrlabs/games@2.10.0
+  - @winrlabs/web3@0.3.0
+  - @winrlabs/web3-games@1.0.0
+
+## 1.0.0
+
+### Minor Changes
+
 - 308fbee: Bet History
 
 ### Patch Changes