+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @winrlabs/web3
 
+## 0.3.0
+
+### Minor Changes
+
+- af62052: Constant url
+
+### Patch Changes
+
+- Updated dependencies [af62052]
+  - @winrlabs/api@1.1.0
+
 ## 0.1.0
 
 ### Minor Changes